+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Alpha 3.9.3
+* Overhauled rails
+  * Added Copper Rails and Golden Rails
 * Critical Hits now deal 25% more damage instead of 50%
 * Fixed timber getting more trees than expected
 * Fixed spiders attack reach and skeletons riding them